feat: change default to infer sass api and load correct functions
Unlike the legacy api, the modern api passes a single array of arguments
to custom functions.
